:truck: chore(api) move user reconciliation service to `src/`
