+++ api/lib/domain/usecases/index.js
@@ -43,6 +43,7 @@ import { config } from '../../../src/shared/config.js';
 import * as codeGenerator from '../../../src/shared/domain/services/code-generator.js';
 import { cryptoService } from '../../../src/shared/domain/services/crypto-service.js';
 import { tokenService } from '../../../src/shared/domain/services/token-service.js';
+import * as userReconciliationService from '../../../src/shared/domain/services/user-reconciliation-service.js';
 import * as userService from '../../../src/shared/domain/services/user-service.js';
 import * as passwordValidator from '../../../src/shared/domain/validators/password-validator.js';
 import * as userValidator from '../../../src/shared/domain/validators/user-validator.js';
@@ -75,7 +76,6 @@ import { certificationCompletedJobRepository } from '../../infrastructure/reposi
 import * as organizationMemberIdentityRepository from '../../infrastructure/repositories/organization-member-identity-repository.js';
 import * as targetProfileShareRepository from '../../infrastructure/repositories/target-profile-share-repository.js';
 import * as learningContentConversionService from '../services/learning-content/learning-content-conversion-service.js';
-import * as userReconciliationService from '../services/user-reconciliation-service.js';
 import * as organizationCreationValidator from '../validators/organization-creation-validator.js';
 import * as organizationValidator from '../validators/organization-with-tags-and-target-profiles-script.js';
