+++ api/src/shared/domain/services/user-reconciliation-service.js
@@ -4,19 +4,19 @@ import fp from 'lodash/fp.js';
 const { pipe } = fp;
 import randomString from 'randomstring';
 
-import { isOneStringCloseEnoughFromMultipleStrings } from '../../../src/evaluation/domain/services/string-comparison-service.js';
+import { isOneStringCloseEnoughFromMultipleStrings } from '../../../evaluation/domain/services/string-comparison-service.js';
 import {
   normalizeAndRemoveAccents,
   removeSpecialCharacters,
-} from '../../../src/evaluation/domain/services/validation-treatments.js';
-import { LEVENSHTEIN_DISTANCE_MAX_RATE, STUDENT_RECONCILIATION_ERRORS } from '../../../src/shared/domain/constants.js';
+} from '../../../evaluation/domain/services/validation-treatments.js';
+import { LEVENSHTEIN_DISTANCE_MAX_RATE, STUDENT_RECONCILIATION_ERRORS } from '../../../shared/domain/constants.js';
 import {
   AlreadyRegisteredUsernameError,
   NotFoundError,
   OrganizationLearnerAlreadyLinkedToInvalidUserError,
   OrganizationLearnerAlreadyLinkedToUserError,
-} from '../../../src/shared/domain/errors.js';
-import { areTwoStringsCloseEnough } from '../../../src/shared/domain/services/string-comparison-service.js';
+} from '../../../shared/domain/errors.js';
+import { areTwoStringsCloseEnough } from '../../../shared/domain/services/string-comparison-service.js';
 
 const STRICT_MATCH_RATIO = 0;
